pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "flask-dance>=7.1.0",
    "flask-login>=0.6.3",
    "gitpython>=3.1.46",
    "oauthlib>=3.3.1",
    "openai>=2.28.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.12.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.42.0",
]
